implement update action on goal resource
-create form in html for user to choose a goal by ID and update
its title and description
-add an event handler for the form submission, update API file using
ajax to make a patch request

+++ assets/scripts/main/events.js
@@ -19,17 +19,28 @@ const onCreateGoal = (event) => {
 }
 
 const onGetGoal = (event) => {
-  const formData = getFormFields(event.target)
   event.preventDefault()
+  const formData = getFormFields(event.target)
   api.getGoal(formData)
     .then(ui.getGoalSuccess)
     .catch(ui.failure)
 }
 
 const onDeleteGoal = (event) => {
-  const id = $(event.target).data('id')
-  api.deleteGoal(id)
-    .then(ui.deleteGoalSuccess)
+  event.preventDefault()
+  if (event.target.matches('.delete-goal')) {
+    const id = $(event.target).data('id')
+    api.deleteGoal(id)
+      .then(ui.deleteGoalSuccess)
+      .catch(ui.failure)
+  }
+}
+
+const onUpdateGoal = (event) => {
+  event.preventDefault()
+  const formData = getFormFields(event.target)
+  api.updateGoal(formData)
+    .then(ui.updateGoalSuccess)
     .catch(ui.failure)
 }
 
@@ -38,6 +49,7 @@ const addHandlers = () => {
   $('.create-goal-form').on('submit', onCreateGoal)
   $('.get-goal-form').on('submit', onGetGoal)
   $('.content').on('click', onDeleteGoal)
+  $('.update-goal-form').on('submit', onUpdateGoal)
 }
 
 module.exports = {
